dashboard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
aiofiles==24.1.0
playwright==1.49.0
apscheduler==3.10.4
aiosqlite==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
sse-starlette==2.1.0
